fix(flow): 分离渲染组件类型并精简E2E测试
